Feature alpine images (#8)
Move to using smaller Alpine images.

We might want to offer other choices out of the box, for example, only using
official repos, or ubuntu's new minimal image.

With that, we have changed the structure to allow different docker image sets
to be selected for use. Different images will need different settings in the
docker-compose.yaml file, for example when mapping folders/files, so
we now have namespaced DockerFiles and docker-compose views.

This should make it easier for folks to customise their experience if they want
to, with less chance of our changes affecting their setup in the future.

With the change to Alpine built images (using the Alpine repos), there are some
compromises in the modules applied for now and the different PHP versions
are constructed through using different Alpine versions.

These compromises are acceptable for us at the moment.